Fix dtype mismatch in class weights for pass policy
The class weights tensor was defaulting to float64 but PyTorch
cross entropy loss expects float32. Added explicit dtype.

+++ scripts/train_pass_policy.py
@@ -72,7 +72,7 @@ class PassPlayDataset(Dataset):
         n_pass = (self.labels == 0).sum()
         n_play = (self.labels == 1).sum()
         total = len(self.labels)
-        return torch.tensor([total / (2 * n_pass), total / (2 * n_play)])
+        return torch.tensor([total / (2 * n_pass), total / (2 * n_play)], dtype=torch.float32)
 
 
 class PassPlayPolicy(nn.Module):
